+++ src/lib/Basic4WebGL/library/descriptors/assetmanager.descriptor.ts
@@ -8,5 +8,11 @@ export const assetmanagerDescriptor: ModuleDescriptor = {
       params: ['name'],
       returns: (p, _self) => `_sb.get(${p.name})`,
     },
+    {
+      name: 'defineRegion',
+      params: ['newName', 'sourceName', 'x', 'y', 'width', 'height'],
+      body: (p, _self) =>
+        `_sb.defineRegion(${p.newName}, ${p.sourceName}, ${p.x}, ${p.y}, ${p.width}, ${p.height})`,
+    },
   ],
 };